+++ baml_language/crates/bridge_cffi/Cargo.toml
@@ -24,7 +24,6 @@ bundle-http = [ "sys_native/bundle-http" ]
 bex_events = { workspace = true }
 bex_events_native = { workspace = true }
 bex_project = { workspace = true }
-bex_vm = { workspace = true }
 bridge_ctypes = { workspace = true }
 sys_native = { workspace = true }
 sys_types = { workspace = true }
